Alterado tipo de retorno do método get_opcode de u8 para u32

+++ src/tokens/transforms/dlb.rs
@@ -94,7 +94,7 @@ impl TokenMethods for Dlb {
 }
 #[cfg(feature = "bytecode")]
 impl BytecodeTokenMethods for Dlb {
-    fn get_opcode(&self) -> u8 {
+    fn get_opcode(&self) -> u32 {
         0x0a
     }
